wip bootstrap theme
The only real blocker to merging this is that it's using the CDN for bootstrap,
tether, and font-awesome. ikiwiki as a matter of policy does not make sites
that rely on offsite resources, so those need to be pulled into the package (or
dependencies of the debian package).

Also, other parts of ikiwiki use jquery 1.6.2, and I guess that would
need to be upgraded for use by bootstrap. I don't know what the jquery
compatability story is like, so worry that upgrading it could break the
parts of ikiwiki that use it.

Also, it's missing a few bits of styles from ikiwiki's
style.css for ikiwiki's login page, the calendar plugin, etc.